Use library API instead of preservation API to get area and disciplines. These will be depricated in preservation API
