+++ src/modules/PlantConfig/http/LibraryApiClient.ts
@@ -6,6 +6,15 @@ import {ErrorResponse, RegisterResponse, TagFunctionResponse} from './LibraryApi
 
 const Settings = require('../../../../settings.json');
 
+export interface AreaResponse {
+    code: string;
+    description: string;
+}
+
+export interface DisciplineResponse {
+    code: string;
+    description: string;
+}
 
 class LibraryApiError extends Error {
 
@@ -126,6 +135,44 @@ class LibraryApiClient extends ApiClient {
                 endpoint,
                 settings
             );
+            return result.data;
+        }
+        catch (error) {
+            throw getLibraryApiError(error);
+        }
+    }
+
+    /**
+     * Get areas
+     *
+     * @param setRequestCanceller Returns a function that can be called to cancel the request
+     */
+    async getAreas(setRequestCanceller?: RequestCanceler): Promise<AreaResponse[]> {
+        const endpoint = '/Areas';
+        const settings: AxiosRequestConfig = {};
+        this.setupRequestCanceler(settings, setRequestCanceller);
+
+        try {
+            const result = await this.client.get<AreaResponse[]>(endpoint, settings);
+            return result.data;
+        }
+        catch (error) {
+            throw getLibraryApiError(error);
+        }
+    }
+
+    /**
+     * Get disciplines
+     *
+     * @param setRequestCanceller Returns a function that can be called to cancel the request
+     */
+    async getDisciplines(setRequestCanceller?: RequestCanceler): Promise<DisciplineResponse[]> {
+        const endpoint = '/Disciplines';
+        const settings: AxiosRequestConfig = {};
+        this.setupRequestCanceler(settings, setRequestCanceller);
+
+        try {
+            const result = await this.client.get<DisciplineResponse[]>(endpoint, settings);
             return result.data;
         }
         catch (error) {
